socialNetworking/requirements.txt
apyori==1.1.2
boto3==1.34.70
python-dotenv==1.0.1
